refactor: client를 authApiClient로 변경한다
