+++ front-end/src/services/index.ts
@@ -1,2 +1,3 @@
-export { default as client } from "./Client";
+// eslint-disable-next-line import/prefer-default-export
+export { default as authApiClient } from "./api/authApiClient";
 export { default as storage } from "./Storage";
